+++ homeassistant/util/yaml/const.py
@@ -1,0 +1,4 @@
+"""Constants."""
+SECRET_YAML = 'secrets.yaml'
+
+_SECRET_NAMESPACE = 'homeassistant'
