+++ tests/common.py
@@ -15,7 +15,7 @@ from io import StringIO
 from unittest.mock import MagicMock, Mock, patch
 
 import homeassistant.util.dt as date_util
-import homeassistant.util.yaml as yaml
+import homeassistant.util.yaml.loader as yaml_loader
 
 from homeassistant import auth, config_entries, core as ha, loader
 from homeassistant.auth import (
@@ -680,7 +680,7 @@ def patch_yaml_files(files_dict, endswith=True):
         # Not found
         raise FileNotFoundError("File not found: {}".format(fname))
 
-    return patch.object(yaml, 'open', mock_open_f, create=True)
+    return patch.object(yaml_loader, 'open', mock_open_f, create=True)
 
 
 def mock_coro(return_value=None, exception=None):
